pyproject.toml
[project]
name = "ai-mediator"
version = "0.1.0"
description = "AI медиатор конфликтов в отношениях"
requires-python = ">=3.11"

# Все зависимости в одном месте
dependencies = [
    # API dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.29.0",
    
    # Bot dependencies  
    "python-telegram-bot>=20.6",
    
    # Shared dependencies
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.3.0",
    "httpx>=0.25.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0", 
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

# Группы зависимостей для uv
api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0", 
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "slowapi>=0.1.9",
]

bot = [
    "python-telegram-bot>=20.6",
    "loguru>=0.7.0",
]

tools = [
    "click>=8.1.0",
    "rich>=13.6.0",
    "jsonschema>=4.19.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared", "services", "tools"]
